MOD: change MiniblockNumber to u64

+++ dal/src/blocks_dal.rs
@@ -124,7 +124,7 @@ impl BlocksDal<'_, '_> {
             .unwrap()
             .number
             .unwrap_or(0);
-        MiniblockNumber(number as u32)
+        MiniblockNumber(number as u64)
     }
 
     #[tracing::instrument(name = "get_newest_l1_batch_header", skip_all)]
@@ -192,8 +192,8 @@ impl BlocksDal<'_, '_> {
         .unwrap();
 
         Some((
-            MiniblockNumber(row.min? as u32),
-            MiniblockNumber(row.max? as u32),
+            MiniblockNumber(row.min? as u64),
+            MiniblockNumber(row.max? as u64),
         ))
     }
